(refactor) split helper search function into its own file away from the primary middleware
